Upgraded Tree Node Removal!
When an object is removed, the parent node checks if it's possible to absorb the leafs' objects without surpassing the limit of objects-per-node, if it can, it does and eliminates the childs. Every parent attempts this so the tree "Unsplits" itself recursively as much as it can, effectively "pruning" the tree from unnecesary empty or almost empty nodes.