Added Workaround for Game Play Save&Load
For some reason when a .FBX meta file exists when initializing the engine, when a scene is reloaded after stopping game the highest objects in the heriarchy are made children of the old root instead of the new one, so I reparent them to the new root in such circumstance.

Until we find a solid fix for the .fbx meta, this prevents the Play/Stop of the game to break.